moved the visibility helper to MapUnitCombatant, where it really should be
